sets up the IterableFlexView view model
