+++ app/build.gradle
@@ -68,6 +68,8 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-intents:3.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.2.0'
     androidTestImplementation 'br.com.concretesolutions:kappuccino:1.2.1'
+
+    //
 }
 tasks.withType(Test) {
     jacoco.includeNoLocationClasses = true
