+++ iterableapi-ui/src/main/java/com/iterable/iterableapi/ui/flex/IterableFlexViewFragment.kt
@@ -1,57 +1,37 @@
 package com.iterable.iterableapi.ui.flex
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
 import com.iterable.iterableapi.*
 import com.iterable.iterableapi.ui.R
 
 class IterableFlexViewFragment : Fragment() {
-    // example flex message data from a payload
-    private val flexMessageMetaData: IterableFlexMessageMetaData = IterableFlexMessageMetaData(
-        "doibjo4590340oidiobnw",
-        "mbn8489b7ehycy",
-        "noj9iyjthfvhs",
-        false
-    )
 
-    private val flexMessageButtons: List<IterableFlexMessageButton> = listOf(
-        IterableFlexMessageButton("reward-button", "REDEEM MEOW", "success")
-    )
-
-    private val flexMessageImages: List<IterableFlexMessageImage> = listOf(
-        IterableFlexMessageImage("coffee-image", "https://example-image-url.com/first-image")
-    )
-
-    private val flexMessageText: List<IterableFlexMessageText> = listOf(
-        IterableFlexMessageText("body", "CATS RULE!!!")
-    )
-
-    private val flexMessageElements: IterableFlexMessageElements = IterableFlexMessageElements(
-        "hero",
-        flexMessageButtons,
-        flexMessageImages,
-        flexMessageText
-    )
-
-    private var flexMessage: IterableFlexMessage = IterableFlexMessage(flexMessageMetaData, flexMessageElements,  "")
+    private lateinit var viewModel: IterableFlexViewViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+        Log.i("ItblFlexViewFragment", "Called ViewModelProviders.of")
+        viewModel = ViewModelProvider(this).get(IterableFlexViewViewModel::class.java)
+
         var view = inflater.inflate(R.layout.iterable_flex_view_fragment, container, false)
 
         var flexMessageText = view.findViewById<TextView>(R.id.flexMessageBody)
         var flexMessageButton = view.findViewById<Button>(R.id.flexMessageButton)
 
-        flexMessageText.text = flexMessage.elements.text[0].text
-        flexMessageButton.text = flexMessage.elements.buttons[0].title
+        flexMessageText.text = viewModel.flexMessage.elements.text[0].text
+        flexMessageButton.text = viewModel.flexMessage.elements.buttons[0].title
 
         return view
     }
