--- iterableapi-ui/src/main/java/com/iterable/iterableapi/ui/flex/IterableFlexViewViewModel.java
@@ -1,4 +1,0 @@
-package com.iterable.iterableapi.ui.flex;
-
-public class IterableFlexViewViewModel {
-}
